bugfix determination of the first backup to retain
sample: 
today = 2022-09-12. bd = 2. old code selects the last backup prior 2022-09-10. the selected backup is 2022-02-09 23:59:59 or older.
the result is that the last backup prior 2022-09-10 will not be deleted. An this this is wrong. change the sql to select the first backup to keep. 
this results in correct deletion of all backups prio 2022-09-10.